+++ dio/test/cancel_token_test.dart
@@ -1,3 +1,5 @@
+import 'dart:typed_data' show Uint8List;
+
 import 'package:dio/dio.dart';
 import 'package:dio/src/adapters/io_adapter.dart';
 import 'package:dio_test/util.dart';
@@ -15,15 +17,13 @@ void main() {
       expectLater(
         token.whenCancel,
         completion(
-          (error) {
-            return error is DioException &&
-                error.type == DioExceptionType.cancel &&
-                error.error == reason;
-          },
+          matchesDioException(DioExceptionType.cancel),
         ),
       );
       token.requestOptions = RequestOptions();
       token.cancel(reason);
+      token.cancel('after cancelled');
+      expect(token.cancelError?.error, reason);
     });
 
     test('cancel without use does not throw (#1765)', () async {
@@ -87,5 +87,97 @@ void main() {
         verify(request.abort()).called(1);
       }
     });
+
+    test('throws if cancelled before making requests', () async {
+      final cancelToken = CancelToken();
+
+      bool walkThroughHandlers = false;
+      final interceptor = QueuedInterceptorsWrapper(
+        onRequest: (options, handler) {
+          walkThroughHandlers = true;
+          handler.next(options);
+        },
+      );
+
+      cancelToken.cancel();
+      final dio = Dio();
+      dio.interceptors.add(interceptor);
+      await expectLater(
+        () => dio.get('/test', cancelToken: cancelToken),
+        throwsDioException(
+          DioExceptionType.cancel,
+          matcher: isA<DioException>(),
+        ),
+      );
+      expect(walkThroughHandlers, isFalse);
+    });
   });
+
+  test(
+    'deallocates HttpClientRequest',
+    () async {
+      final client = MockHttpClient();
+      final dio = Dio();
+      dio.httpClientAdapter = IOHttpClientAdapter(
+        createHttpClient: () => client,
+      );
+      final token = CancelToken();
+      final requests = <MockHttpClientRequest>{};
+      final requestsReferences = <WeakReference<MockHttpClientRequest>>{};
+      when(client.openUrl(any, any)).thenAnswer((_) async {
+        final request = MockHttpClientRequest();
+        requests.add(request);
+        requestsReferences.add(WeakReference(request));
+        when(request.close()).thenAnswer((_) async {
+          final response = MockHttpClientResponse();
+          when(response.headers).thenReturn(MockHttpHeaders());
+          when(response.statusCode).thenReturn(200);
+          when(response.reasonPhrase).thenReturn('OK');
+          when(response.isRedirect).thenReturn(false);
+          when(response.redirects).thenReturn([]);
+          when(response.cast())
+              .thenAnswer((_) => const Stream<Uint8List>.empty());
+          await Future.delayed(const Duration(milliseconds: 200));
+          return response;
+        });
+        when(request.abort()).thenAnswer((realInvocation) {
+          requests.remove(request);
+        });
+        return request;
+      });
+
+      final futures = [
+        dio.get('https://does.not.exists', cancelToken: token),
+        dio.get('https://does.not.exists', cancelToken: token),
+      ];
+      for (final future in futures) {
+        expectLater(
+          future,
+          throwsDioException(DioExceptionType.cancel),
+        );
+      }
+
+      // Opening requests.
+      await Future.delayed(const Duration(milliseconds: 100));
+      token.cancel();
+      // Aborting requests.
+      await Future.delayed(const Duration(seconds: 1));
+      expect(requests, isEmpty);
+
+      try {
+        await Future.wait(futures);
+      } catch (_) {
+        // Waiting here until all futures are completed.
+      }
+      expect(requests, isEmpty);
+      expect(requestsReferences, hasLength(2));
+
+      // GC.
+      produceGarbage();
+      await Future.delayed(const Duration(seconds: 1));
+      expect(requestsReferences.every((e) => e.target == null), isTrue);
+    },
+    tags: ['gc'],
+    testOn: 'vm',
+  );
 }