+++ dio_test/lib/src/test/cors_tests.dart
@@ -1,6 +1,7 @@
 import 'package:dio/dio.dart';
-import 'package:dio_test/util.dart';
 import 'package:test/test.dart';
+
+import '../../util.dart';
 
 /// Test that browsers can correctly classify requests as
 /// either "simple" or "preflighted". Reference:
@@ -77,7 +78,7 @@ void corsTests(
               options: Options(
                 validateStatus: (status) => true,
                 contentType: Headers.textPlainContentType,
-                sendTimeout: Duration(seconds: 1),
+                sendTimeout: const Duration(seconds: 1),
               ),
             );
           },