+++ dio_test/lib/src/test/download_tests.dart
@@ -3,9 +3,10 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:dio/dio.dart';
-import 'package:dio_test/util.dart';
 import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
+
+import '../../util.dart';
 
 void downloadTests(
   Dio Function(String baseUrl) create,
@@ -64,10 +65,13 @@ void downloadTests(
           cancelToken.cancel();
         });
 
-        final completer = Completer();
-        res.data!.stream.listen((event) {}, onError: (e, s) {
-          completer.completeError(e, s);
-        });
+        final completer = Completer<Never>();
+        res.data!.stream.listen(
+          (event) {},
+          onError: (e, s) {
+            completer.completeError(e, s);
+          },
+        );
 
         await expectLater(
           completer.future,
@@ -246,7 +250,7 @@ void downloadTests(
           dio.download(
             '/drip?delay=0&duration=6&numbytes=3',
             p.join(tmp.path, 'download_timeout.md'),
-            options: Options(receiveTimeout: Duration(seconds: 1)),
+            options: Options(receiveTimeout: const Duration(seconds: 1)),
           ),
           throwsDioException(
             DioExceptionType.receiveTimeout,
@@ -262,7 +266,7 @@ void downloadTests(
             dio.download(
               '/drip?delay=0&duration=6&numbytes=12',
               p.join(tmp.path, 'download_timeout.md'),
-              options: Options(receiveTimeout: Duration(seconds: 1)),
+              options: Options(receiveTimeout: const Duration(seconds: 1)),
             ),
             completes,
           );
@@ -375,6 +379,55 @@ void downloadTests(
           ),
           completes,
         );
+      });
+
+      test('append bytes to previous download', () async {
+        final cancelToken = CancelToken();
+        final path = p.join(tmp.path, 'download_3.txt');
+        final requestedBytes = 1024 * 1024 * 10;
+        int recievedBytes1 = 0;
+        await expectLater(
+          dio.download(
+            '/bytes/$requestedBytes',
+            path,
+            cancelToken: cancelToken,
+            onReceiveProgress: (c, t) {
+              if (c > 5000) {
+                recievedBytes1 = c;
+                cancelToken.cancel();
+              }
+            },
+            deleteOnError: false,
+          ),
+          throwsDioException(
+            DioExceptionType.cancel,
+            stackTraceContains: 'test/download_tests.dart',
+          ),
+        );
+
+        final cancelToken2 = CancelToken();
+        int recievedBytes2 = 0;
+        expectLater(
+          dio.download(
+            '/bytes/$requestedBytes',
+            path,
+            cancelToken: cancelToken,
+            onReceiveProgress: (c, t) {
+              recievedBytes2 = c;
+              if (c > 5000) {
+                cancelToken2.cancel();
+              }
+            },
+            deleteOnError: false,
+            fileAccessMode: FileAccessMode.append,
+          ),
+          throwsDioException(
+            DioExceptionType.cancel,
+            stackTraceContains: 'test/download_tests.dart',
+          ),
+        );
+        await Future.delayed(const Duration(milliseconds: 100), () {});
+        expect(File(path).lengthSync(), recievedBytes1 + recievedBytes2);
       });
     },
     testOn: 'vm',