+++ plugins/native_dio_adapter/lib/src/cronet_adapter.dart
@@ -1,4 +1,4 @@
-import 'dart:typed_data';
+import 'dart:typed_data' show Uint8List;
 
 import 'package:cronet_http/cronet_http.dart';
 import 'package:dio/dio.dart';
@@ -8,14 +8,19 @@ import 'conversion_layer_adapter.dart';
 /// to the native platform by making use of
 /// [cronet_http](https://pub.dev/packages/cronet_http).
 class CronetAdapter implements HttpClientAdapter {
-  CronetAdapter(CronetEngine? engine)
-      : _conversionLayer = ConversionLayerAdapter(
+  CronetAdapter(
+    CronetEngine? engine, {
+    bool closeEngine = true,
+  }) : _conversionLayer = ConversionLayerAdapter(
           engine == null
               ? CronetClient.defaultCronetEngine()
-              : CronetClient.fromCronetEngine(engine),
+              : CronetClient.fromCronetEngine(engine, closeEngine: closeEngine),
         );
 
   final ConversionLayerAdapter _conversionLayer;
+
+  /// The underlying conversion layer adapter.
+  ConversionLayerAdapter get adapter => _conversionLayer;
 
   @override
   void close({bool force = false}) => _conversionLayer.close(force: force);