+++ plugins/native_dio_adapter/test/conversion_layer_adapter_test.dart
@@ -1,9 +1,9 @@
 import 'dart:typed_data';
 
 import 'package:dio/dio.dart';
+import 'package:flutter_test/flutter_test.dart';
 import 'package:http/http.dart';
 import 'package:native_dio_adapter/src/conversion_layer_adapter.dart';
-import 'package:flutter_test/flutter_test.dart';
 
 import 'client_mock.dart';
 
@@ -27,12 +27,13 @@ void main() {
   });
 
   test('headers', () async {
-    final mock = ClientMock()..response = StreamedResponse(Stream.empty(), 200);
+    final mock = ClientMock()
+      ..response = StreamedResponse(const Stream.empty(), 200);
     final cla = ConversionLayerAdapter(mock);
 
     await cla.fetch(
       RequestOptions(path: '', headers: {'foo': 'bar'}),
-      Stream.empty(),
+      const Stream.empty(),
       null,
     );
 
@@ -42,14 +43,15 @@ void main() {
   test('download stream', () async {
     final mock = ClientMock()
       ..response = StreamedResponse(
-          Stream.fromIterable(<Uint8List>[
-            Uint8List.fromList([10, 1]),
-            Uint8List.fromList([1, 4]),
-            Uint8List.fromList([5, 1]),
-            Uint8List.fromList([1, 1]),
-            Uint8List.fromList([2, 4]),
-          ]),
-          200);
+        Stream.fromIterable(<Uint8List>[
+          Uint8List.fromList([10, 1]),
+          Uint8List.fromList([1, 4]),
+          Uint8List.fromList([5, 1]),
+          Uint8List.fromList([1, 1]),
+          Uint8List.fromList([2, 4]),
+        ]),
+        200,
+      );
     final cla = ConversionLayerAdapter(mock);
 
     final resp = await cla.fetch(
@@ -59,5 +61,53 @@ void main() {
     );
 
     expect(await resp.stream.length, 5);
+  });
+
+  test('request cancellation', () {
+    final mock = AbortClientMock();
+    final cla = ConversionLayerAdapter(mock);
+    final cancelToken = CancelToken();
+
+    Future<void>.delayed(const Duration(seconds: 1)).then(
+      (value) {
+        cancelToken.cancel();
+      },
+    );
+
+    expectLater(
+      () => cla.fetch(
+        RequestOptions(path: ''),
+        null,
+        cancelToken.whenCancel,
+      ),
+      throwsA(isA<AbortedError>()),
+    );
+  });
+
+  test('request cancellation with Dio', () async {
+    final mock = AbortClientMock();
+    final cla = ConversionLayerAdapter(mock);
+    final dio = Dio();
+    dio.httpClientAdapter = cla;
+
+    final cancelToken = CancelToken();
+
+    Future<void>.delayed(const Duration(seconds: 1)).then(
+      (value) {
+        cancelToken.cancel();
+      },
+    );
+
+    await expectLater(
+      () => dio.get<ResponseBody>('', cancelToken: cancelToken),
+      throwsA(
+        isA<DioException>().having(
+          (e) => e.type,
+          'type',
+          DioExceptionType.cancel,
+        ),
+      ),
+    );
+    expect(mock.isRequestCanceled, true);
   });
 }